Add files via upload
Text Simplification (TS) is essentially a task in natural language generation, wherein the produced text exhibits decreased complexity in both vocabulary and sentence structure. Despite these simplifications, the goal is to retain the original information and meaning intact.
TS serves as a valuable preprocessing step to enhance the outcomes of various NLP tasks. Examples include parsing, information extraction, question generation, and text summarization.
We used T5 model, which is a sequence-to-sequence Transformer-based model.